tar driver finished!

+++ include/klib.h
@@ -11,7 +11,7 @@ size_t kstrlen(const char *);
 size_t kprintf(const char *, ...);
 char *kstrchr(char *, char);
 char *kstrrchr(char *, char);
-size_t kstrcmp(char *, char *);
+int kstrcmp(char *, char *);
 char *kstrstr(char *, char *);
 void *kmemcpy(void *, void *, size_t);
 void *kmemset(void *, int, size_t);
